refactor(value): split types.rs into cohesive modules (§7-8)
Split the 941-line value/types.rs (over the 500-line limit) into four
modules under src/value/, all under 500 lines:

- types.rs (80)         - what_type_name + allomorph_type_name free fns
- types_eqv.rs (347)    - Value::eqv (with nested allomorph_kind)
- types_truthy.rs (144) - is_failure_handled / mark_failure_handled / truthy
- types_isa.rs (379)    - Value::isa_check + does_check

All moved methods are pub(crate) on Value, so the split is relocation
only (no visibility bumps). types_eqv/types_isa import allomorph_type_name
from the main module. The fn multiset is byte-identical. No behavior change.
