+++ src/value/types.rs
@@ -1,866 +1,5 @@
 use super::*;
 
-impl Value {
-    /// Type-strict structural equivalence (Raku `eqv` operator).
-    /// See raku-doc: Language/operators.rakudoc "infix eqv"
-    ///
-    /// Returns True if two arguments are structurally the same, i.e. from the
-    /// same type and (recursively) contain equivalent values.
-    /// Unlike PartialEq (used for `==`), this does NOT allow cross-type comparisons:
-    ///   1 eqv 1.0  → False  (Int vs Num)
-    ///   [1,2] eqv (1,2)  → False  (Array vs List)
-    pub(crate) fn eqv(&self, other: &Self) -> bool {
-        // Unwrap Scalar/ContainerRef containers: eqv looks through containerization
-        if let Value::Scalar(inner) = self {
-            return inner.eqv(other);
-        }
-        if let Value::Scalar(inner) = other {
-            return self.eqv(inner);
-        }
-        // Deref to an OWNED clone (releasing the cell lock) before recursing:
-        // when both sides alias the SAME cell (e.g. two pairs built from the same
-        // `key => $var`), holding the lock across the recursive `eqv` would lock
-        // the same non-reentrant Mutex twice and deadlock.
-        if matches!(self, Value::ContainerRef(_)) {
-            return self.deref_container().eqv(other);
-        }
-        if matches!(other, Value::ContainerRef(_)) {
-            return self.eqv(&other.deref_container());
-        }
-        // Junction threading: if either side is a junction, thread eqv
-        // through it and return the boolean result of the junction.
-        if let Value::Junction { kind, values } = other {
-            let results: Vec<bool> = values.iter().map(|v| self.eqv(v)).collect();
-            return match kind {
-                crate::value::JunctionKind::Any => results.iter().any(|&b| b),
-                crate::value::JunctionKind::All => results.iter().all(|&b| b),
-                crate::value::JunctionKind::One => results.iter().filter(|&&b| b).count() == 1,
-                crate::value::JunctionKind::None => results.iter().all(|&b| !b),
-            };
-        }
-        if let Value::Junction { kind, values } = self {
-            let results: Vec<bool> = values.iter().map(|v| v.eqv(other)).collect();
-            return match kind {
-                crate::value::JunctionKind::Any => results.iter().any(|&b| b),
-                crate::value::JunctionKind::All => results.iter().all(|&b| b),
-                crate::value::JunctionKind::One => results.iter().filter(|&&b| b).count() == 1,
-                crate::value::JunctionKind::None => results.iter().all(|&b| !b),
-            };
-        }
-        match (self, other) {
-            // Arrays/Lists: must be same container type (Array vs List) and recursively eqv
-            // eqv ignores Scalar wrapping — only Array vs List distinction matters
-            (Value::Array(a, a_kind), Value::Array(b, b_kind)) => {
-                a_kind.is_real_array() == b_kind.is_real_array()
-                    && a.len() == b.len()
-                    && a.iter().zip(b.iter()).all(|(x, y)| x.eqv(y))
-            }
-            // Hashes: recursively use eqv for values
-            (Value::Hash(a), Value::Hash(b)) => {
-                a.len() == b.len() && a.iter().all(|(k, v)| b.get(k).is_some_and(|bv| v.eqv(bv)))
-            }
-            // Pairs: recursively use eqv for values (Pair and ValuePair are equivalent)
-            (Value::Pair(ak, av), Value::Pair(bk, bv)) => ak == bk && av.eqv(bv),
-            (Value::ValuePair(ak, av), Value::ValuePair(bk, bv)) => ak.eqv(bk) && av.eqv(bv),
-            (Value::Pair(ak, av), Value::ValuePair(bk, bv)) => {
-                matches!(bk.as_ref(), Value::Str(s) if s.as_str() == ak) && av.eqv(bv)
-            }
-            (Value::ValuePair(ak, av), Value::Pair(bk, bv)) => {
-                matches!(ak.as_ref(), Value::Str(s) if s.as_str() == bk) && av.eqv(bv)
-            }
-            // Captures: recursively use eqv for positional and named elements
-            (
-                Value::Capture {
-                    positional: ap,
-                    named: an,
-                },
-                Value::Capture {
-                    positional: bp,
-                    named: bn,
-                },
-            ) => {
-                ap.len() == bp.len()
-                    && ap.iter().zip(bp.iter()).all(|(x, y)| x.eqv(y))
-                    && an.len() == bn.len()
-                    && an
-                        .iter()
-                        .all(|(k, v)| bn.get(k).is_some_and(|bv| v.eqv(bv)))
-            }
-            // Slips: recursively use eqv for elements
-            (Value::Slip(a), Value::Slip(b)) => {
-                a.len() == b.len() && a.iter().zip(b.iter()).all(|(x, y)| x.eqv(y))
-            }
-            // Seqs: recursively use eqv for elements
-            (Value::Seq(a), Value::Seq(b)) => {
-                a.len() == b.len() && a.iter().zip(b.iter()).all(|(x, y)| x.eqv(y))
-            }
-            // Num: use bit-exact comparison to distinguish signed zeros,
-            // but treat all NaN bit patterns as identical (Raku considers all NaN equal).
-            (Value::Num(a), Value::Num(b)) => {
-                if a.is_nan() && b.is_nan() {
-                    true
-                } else {
-                    a.to_bits() == b.to_bits()
-                }
-            }
-            // Complex: use bit-exact comparison for both components,
-            // treating all NaN bit patterns as identical.
-            (Value::Complex(ar, ai), Value::Complex(br, bi)) => {
-                let re_eq = if ar.is_nan() && br.is_nan() {
-                    true
-                } else {
-                    ar.to_bits() == br.to_bits()
-                };
-                let im_eq = if ai.is_nan() && bi.is_nan() {
-                    true
-                } else {
-                    ai.to_bits() == bi.to_bits()
-                };
-                re_eq && im_eq
-            }
-            // Same-type scalar comparisons delegate to PartialEq
-            // BigInt: both sides BigInt — use PartialEq
-            (Value::BigInt(_), Value::BigInt(_)) => self == other,
-            // Cross-representation Int/BigInt: compare numerically
-            (Value::Int(a), Value::BigInt(b)) => NumBigInt::from(*a) == **b,
-            (Value::BigInt(a), Value::Int(b)) => **a == NumBigInt::from(*b),
-            // Rat/FatRat: structural equality (n == n, d == d), including NaN (0/0)
-            (Value::Rat(n1, d1), Value::Rat(n2, d2))
-            | (Value::FatRat(n1, d1), Value::FatRat(n2, d2)) => n1 == n2 && d1 == d2,
-            // Sets: eqv must distinguish elements that share a string key but
-            // differ in type — specifically an allomorph (e.g. the IntStr <42>)
-            // from a plain value (Int 42), which rakudo separates by `.WHICH`.
-            // We compare the allomorph kind of each typed element rather than a
-            // full type-strict eqv, because mutsu does not always retain a Set
-            // element's exact numeric type (a Rat element can fall back to its
-            // Str key), and a full eqv would wrongly split two equal Rat sets.
-            // Mutability (Set vs SetHash) is part of the type, so eqv must
-            // distinguish them (`Set.new(42) eqv SetHash.new(42)` is False).
-            // Raku's set operators (`(|)`/`(&)` etc.) always yield an immutable
-            // Set regardless of operand mutability, so comparing the flag here
-            // matches values produced by those operators too.
-            (Value::Set(a, a_mut), Value::Set(b, b_mut)) => {
-                fn allomorph_kind(v: &Value) -> Option<String> {
-                    match v {
-                        Value::Mixin(inner, mixins) => allomorph_type_name(inner, mixins),
-                        _ => None,
-                    }
-                }
-                a_mut == b_mut
-                    && a.elements.len() == b.elements.len()
-                    && a.elements.iter().all(|k| {
-                        b.elements.contains(k)
-                            && allomorph_kind(&a.typed_key(k)) == allomorph_kind(&b.typed_key(k))
-                    })
-            }
-            // Bag/Mix: like Set, eqv distinguishes the immutable variant from
-            // its mutable QuantHash (Bag vs BagHash, Mix vs MixHash). The data
-            // comparison is delegated to PartialEq, which ignores the flag.
-            (Value::Bag(a, a_mut), Value::Bag(b, b_mut)) => a_mut == b_mut && a == b,
-            (Value::Mix(a, a_mut), Value::Mix(b, b_mut)) => a_mut == b_mut && a == b,
-            (Value::Int(_), Value::Int(_))
-            | (Value::Str(_), Value::Str(_))
-            | (Value::Bool(_), Value::Bool(_))
-            | (Value::BigRat(_, _), Value::BigRat(_, _))
-            | (Value::Enum { .. }, Value::Enum { .. })
-            | (Value::Regex(_), Value::Regex(_))
-            | (Value::RegexWithAdverbs { .. }, Value::RegexWithAdverbs { .. })
-            | (Value::Routine { .. }, Value::Routine { .. }) => self == other,
-            (Value::Sub(a), Value::Sub(b)) => {
-                if Arc::ptr_eq(a, b) {
-                    return true;
-                }
-                let a_name = a.name.resolve();
-                let b_name = b.name.resolve();
-                !a_name.is_empty() && a_name == b_name && a.package == b.package
-            }
-            // Signature instances: compare by .raku string (structural equality)
-            (
-                Value::Instance {
-                    class_name: cn_a, ..
-                },
-                Value::Instance {
-                    class_name: cn_b, ..
-                },
-            ) if cn_a == "Signature" && cn_b == "Signature" => {
-                let raku_a = if let Value::Instance { attributes, .. } = self {
-                    attributes.as_map().get("raku").map(|v| v.to_string_value())
-                } else {
-                    None
-                };
-                let raku_b = if let Value::Instance { attributes, .. } = other {
-                    attributes.as_map().get("raku").map(|v| v.to_string_value())
-                } else {
-                    None
-                };
-                raku_a == raku_b
-            }
-            (
-                Value::Instance {
-                    class_name: cn_a,
-                    attributes: a_attrs,
-                    ..
-                },
-                Value::Instance {
-                    class_name: cn_b,
-                    attributes: b_attrs,
-                    ..
-                },
-            ) if cn_a == cn_b
-                && a_attrs.contains_key("year")
-                && a_attrs.contains_key("month")
-                && a_attrs.contains_key("day")
-                && a_attrs.contains_key("hour")
-                && a_attrs.contains_key("minute")
-                && a_attrs.contains_key("second")
-                && a_attrs.contains_key("timezone")
-                && b_attrs.contains_key("year")
-                && b_attrs.contains_key("month")
-                && b_attrs.contains_key("day")
-                && b_attrs.contains_key("hour")
-                && b_attrs.contains_key("minute")
-                && b_attrs.contains_key("second")
-                && b_attrs.contains_key("timezone") =>
-            {
-                let (ay, am, ad, ah, amin, asec, atz) =
-                    crate::builtins::methods_0arg::temporal::datetime_attrs(&(a_attrs).as_map());
-                let (by, bm, bd, bh, bmin, bsec, btz) =
-                    crate::builtins::methods_0arg::temporal::datetime_attrs(&(b_attrs).as_map());
-                ay == by
-                    && am == bm
-                    && ad == bd
-                    && ah == bh
-                    && amin == bmin
-                    && atz == btz
-                    && (asec - bsec).abs() < 1e-6
-            }
-            // Date instances: compare only year/month/day (ignore formatter attrs)
-            (
-                Value::Instance {
-                    class_name: cn_a,
-                    attributes: a_attrs,
-                    ..
-                },
-                Value::Instance {
-                    class_name: cn_b,
-                    attributes: b_attrs,
-                    ..
-                },
-            ) if cn_a == cn_b
-                && a_attrs.contains_key("year")
-                && a_attrs.contains_key("month")
-                && a_attrs.contains_key("day")
-                && !a_attrs.contains_key("hour")
-                && b_attrs.contains_key("year")
-                && b_attrs.contains_key("month")
-                && b_attrs.contains_key("day")
-                && !b_attrs.contains_key("hour") =>
-            {
-                let (ay, am, ad) =
-                    crate::builtins::methods_0arg::temporal::date_attrs(&(a_attrs).as_map());
-                let (by, bm, bd) =
-                    crate::builtins::methods_0arg::temporal::date_attrs(&(b_attrs).as_map());
-                ay == by && am == bm && ad == bd
-            }
-            // StrDistance instances: structural equality on before/after
-            (
-                Value::Instance {
-                    class_name: cn_a,
-                    attributes: a_attrs,
-                    ..
-                },
-                Value::Instance {
-                    class_name: cn_b,
-                    attributes: b_attrs,
-                    ..
-                },
-            ) if cn_a == "StrDistance" && cn_b == "StrDistance" => {
-                let before_eq = match (
-                    a_attrs.as_map().get("before"),
-                    b_attrs.as_map().get("before"),
-                ) {
-                    (Some(a), Some(b)) => a.eqv(b),
-                    (None, None) => true,
-                    _ => false,
-                };
-                let after_eq = match (a_attrs.as_map().get("after"), b_attrs.as_map().get("after"))
-                {
-                    (Some(a), Some(b)) => a.eqv(b),
-                    (None, None) => true,
-                    _ => false,
-                };
-                before_eq && after_eq
-            }
-            // Other Instance types: use identity comparison
-            (Value::Instance { .. }, Value::Instance { .. }) => self == other,
-            // Nil and Package("Any") are eqv: both represent the undefined type object Any.
-            // This matters for containerized contexts (e.g. hash values).
-            (Value::Nil, Value::Package(name)) | (Value::Package(name), Value::Nil)
-                if name == "Any" =>
-            {
-                true
-            }
-            (Value::Range(_, _), Value::Range(_, _))
-            | (Value::RangeExcl(_, _), Value::RangeExcl(_, _))
-            | (Value::RangeExclStart(_, _), Value::RangeExclStart(_, _))
-            | (Value::RangeExclBoth(_, _), Value::RangeExclBoth(_, _))
-            | (Value::GenericRange { .. }, Value::GenericRange { .. })
-            | (Value::LazyList(_), Value::LazyList(_))
-            | (Value::Version { .. }, Value::Version { .. })
-            | (Value::Nil, Value::Nil)
-            | (Value::Package(_), Value::Package(_))
-            | (Value::CompUnitDepSpec { .. }, Value::CompUnitDepSpec { .. })
-            | (Value::Junction { .. }, Value::Junction { .. })
-            | (Value::Promise(_), Value::Promise(_))
-            | (Value::Channel(_), Value::Channel(_))
-            | (Value::Uni { .. }, Value::Uni { .. }) => self == other,
-            // Mixin with only the read-only topic marker is transparent
-            // (used by `with literal { ... }` to flag immutable $_).
-            (Value::Mixin(inner, mix), other)
-                if mix.len() == 1 && mix.contains_key("__mutsu_topic_ro__") =>
-            {
-                inner.eqv(other)
-            }
-            (other, Value::Mixin(inner, mix))
-                if mix.len() == 1 && mix.contains_key("__mutsu_topic_ro__") =>
-            {
-                other.eqv(inner)
-            }
-            // Mixin (allomorphs): compare both base values and mixin maps with eqv
-            (Value::Mixin(a, a_mix), Value::Mixin(b, b_mix)) => {
-                if !a.eqv(b) {
-                    return false;
-                }
-                // Compare mixin maps (e.g. Str part of allomorphs)
-                if a_mix.len() != b_mix.len() {
-                    return false;
-                }
-                a_mix
-                    .iter()
-                    .all(|(k, v)| b_mix.get(k).is_some_and(|bv| v.eqv(bv)))
-            }
-            // Cross-type comparisons always return false for eqv
-            _ => false,
-        }
-    }
-
-    /// Check if this Failure instance is handled.
-    /// Returns true if handled, false otherwise.
-    /// Panics if called on a non-Failure.
-    pub(crate) fn is_failure_handled(&self) -> bool {
-        if let Value::Instance {
-            class_name,
-            attributes,
-            id,
-            ..
-        } = self
-            && class_name.resolve() == "Failure"
-        {
-            // Check the global registry first (shared across clones)
-            if let Some(handled) = super::is_failure_handled(*id) {
-                return handled;
-            }
-            // Fall back to the attribute
-            attributes
-                .as_map()
-                .get("handled")
-                .is_some_and(|v| v.truthy())
-        } else {
-            false
-        }
-    }
-
-    /// Mark this Failure as handled in the global registry.
-    pub(crate) fn mark_failure_handled(&self) {
-        if let Value::Instance { id, .. } = self {
-            super::mark_failure_handled(*id);
-        }
-    }
-
-    pub(crate) fn truthy(&self) -> bool {
-        match self {
-            Value::Bool(b) => *b,
-            Value::Int(i) => *i != 0,
-            Value::BigInt(n) => !n.is_zero(),
-            Value::Num(f) => *f != 0.0 || f.is_nan(),
-            Value::Str(s) => !s.is_empty(),
-            Value::Range(_, _) => true,
-            Value::RangeExcl(_, _) => true,
-            Value::RangeExclStart(_, _) => true,
-            Value::RangeExclBoth(_, _) => true,
-            Value::GenericRange { .. } => true,
-            Value::Array(items, ..) => !items.is_empty(),
-            Value::Hash(items) => !items.is_empty(),
-            Value::Rat(n, _) => *n != 0,
-            Value::FatRat(n, _) => !n.is_zero(),
-            Value::BigRat(n, _) => !n.is_zero(),
-            Value::Complex(r, i) => *r != 0.0 || *i != 0.0,
-            Value::Set(s, _) => !s.is_empty(),
-            Value::Bag(b, _) => !b.is_empty(),
-            Value::Mix(m, _) => !m.is_empty(),
-            Value::Pair(_, _) | Value::ValuePair(_, _) => true,
-            Value::Enum { value, .. } => match value {
-                EnumValue::Int(i) => *i != 0,
-                EnumValue::Str(s) => !s.is_empty(),
-                EnumValue::Generic(v) => v.as_ref().truthy(),
-            },
-            Value::CompUnitDepSpec { .. } => true,
-            Value::Package(_) | Value::ParametricRole { .. } => false,
-            Value::Routine { .. } => true,
-            Value::Sub(_) | Value::WeakSub(_) => true,
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } => {
-                if class_name == "Proc" {
-                    // Proc is truthy when exitcode == 0
-                    return match attributes.as_map().get("exitcode") {
-                        Some(Value::Int(code)) => *code == 0,
-                        _ => false,
-                    };
-                }
-                if class_name == "Failure" {
-                    return false;
-                }
-                // Buf/Blob: truthy when non-empty
-                let cn = class_name.resolve();
-                if cn == "Buf"
-                    || cn == "Blob"
-                    || cn.starts_with("Buf[")
-                    || cn.starts_with("Blob[")
-                    || cn.starts_with("buf")
-                    || cn.starts_with("blob")
-                {
-                    return match attributes.as_map().get("bytes") {
-                        Some(Value::Array(items, ..)) => !items.is_empty(),
-                        _ => false,
-                    };
-                }
-                true
-            }
-            Value::Junction { kind, values } => match kind {
-                JunctionKind::Any => values.iter().any(|v| v.truthy()),
-                JunctionKind::All => values.iter().all(|v| v.truthy()),
-                JunctionKind::One => values.iter().filter(|v| v.truthy()).count() == 1,
-                JunctionKind::None => values.iter().all(|v| !v.truthy()),
-            },
-            Value::Slip(items) => !items.is_empty(),
-            Value::Seq(items) | Value::HyperSeq(items) | Value::RaceSeq(items) => !items.is_empty(),
-            Value::LazyList(_) => true,
-            Value::Promise(p) => p.is_resolved(),
-            Value::Channel(_) => true,
-            Value::Regex(_) | Value::RegexWithAdverbs { .. } => true,
-            Value::Version { .. } => true,
-            Value::Nil => false,
-            Value::Whatever => true,
-            Value::HyperWhatever => true,
-            Value::Capture { positional, named } => !positional.is_empty() || !named.is_empty(),
-            Value::Uni(u) => !u.text.is_empty(),
-            Value::Mixin(inner, mixins) => {
-                if let Some(bool_val) = mixins.get("Bool") {
-                    bool_val.truthy()
-                } else {
-                    inner.truthy()
-                }
-            }
-            Value::Proxy { .. } => true,
-            Value::CustomType { .. } => false,
-            Value::CustomTypeInstance(_) => true,
-            Value::Scalar(inner) => inner.truthy(),
-            Value::ContainerRef(_) => self.with_deref(Value::truthy),
-            Value::LazyThunk(thunk_data) => {
-                let cache = thunk_data.cache.lock().unwrap();
-                if let Some(ref cached) = *cache {
-                    cached.truthy()
-                } else {
-                    // Unforced thunk is truthy (it exists)
-                    true
-                }
-            }
-            Value::LazyIoLines { .. } => true,
-            // An unmaterialized deferred bind token is undefined (and hence
-            // falsy) until written through — see `value_is_defined`.
-            Value::HashEntryRef { .. } => false,
-        }
-    }
-
-    /// Check if this value is an instance of the given type name (Raku `isa` operator).
-    pub(crate) fn isa_check(&self, type_name: &str) -> bool {
-        // For Instance/Package, extract name as owned String for later comparison
-        let owned_name: Option<String> = match self {
-            Value::Instance { class_name, .. } => Some(class_name.resolve()),
-            Value::Package(name) => Some(name.resolve()),
-            _ => None,
-        };
-        let my_type = match self {
-            Value::Int(_) | Value::BigInt(_) => "Int",
-            Value::Num(_) => "Num",
-            Value::Str(_) => "Str",
-            Value::Bool(_) => "Bool",
-            Value::Rat(_, _) => "Rat",
-            Value::FatRat(_, _) => "FatRat",
-            Value::BigRat(_, _) => "Rat",
-            Value::Complex(_, _) => "Complex",
-            Value::Array(..) | Value::LazyList(_) => "Array",
-            Value::Seq(_) => "Seq",
-            Value::HyperSeq(_) => "HyperSeq",
-            Value::RaceSeq(_) => "RaceSeq",
-            Value::Hash(_) => "Hash",
-            Value::Set(_, is_mutable) => {
-                if *is_mutable {
-                    "SetHash"
-                } else {
-                    "Set"
-                }
-            }
-            Value::Bag(_, is_mutable) => {
-                if *is_mutable {
-                    "BagHash"
-                } else {
-                    "Bag"
-                }
-            }
-            Value::Mix(_, is_mutable) => {
-                if *is_mutable {
-                    "MixHash"
-                } else {
-                    "Mix"
-                }
-            }
-            Value::Pair(_, _) | Value::ValuePair(_, _) => "Pair",
-            Value::Range(_, _)
-            | Value::RangeExcl(_, _)
-            | Value::RangeExclStart(_, _)
-            | Value::RangeExclBoth(_, _)
-            | Value::GenericRange { .. } => "Range",
-            Value::Nil => "Nil",
-            Value::Instance { .. } | Value::Package(_) => owned_name.as_deref().unwrap(),
-            Value::Enum { enum_type, .. } => {
-                return enum_type.resolve() == type_name;
-            }
-            Value::Sub(data) => match data.env.get("__mutsu_callable_type") {
-                Some(Value::Str(kind)) if kind.as_str() == "Method" => "Method",
-                Some(Value::Str(kind)) if kind.as_str() == "Submethod" => "Submethod",
-                Some(Value::Str(kind)) if kind.as_str() == "WhateverCode" => "WhateverCode",
-                _ => "Sub",
-            },
-            Value::WeakSub(_) => "Sub",
-            Value::Routine {
-                is_regex: false, ..
-            } => "Sub",
-            Value::Regex(_)
-            | Value::RegexWithAdverbs { .. }
-            | Value::Routine { is_regex: true, .. } => "Regex",
-            Value::Junction { .. } => "Junction",
-            Value::Version { .. } => "Version",
-            Value::Slip(_) => "Slip",
-            Value::Promise(p) => {
-                let cn = p.class_name();
-                if cn != "Promise" && cn == type_name {
-                    return true;
-                }
-                // Also check if type_name is "Promise" (parent)
-                if type_name == "Promise" {
-                    return true;
-                }
-                "Promise"
-            }
-            Value::Channel(_) => "Channel",
-            Value::CompUnitDepSpec { .. } => "CompUnit::DependencySpecification",
-            Value::Whatever => "Whatever",
-            Value::HyperWhatever => "HyperWhatever",
-            Value::Capture { .. } => "Capture",
-            Value::Uni(u) => u.form.as_str(),
-            Value::Mixin(inner, mixins) => {
-                // Check allomorphic type names (IntStr, NumStr, RatStr, ComplexStr, Allomorph)
-                if matches!(
-                    type_name,
-                    "IntStr" | "NumStr" | "RatStr" | "ComplexStr" | "Allomorph"
-                ) {
-                    if let Some(allo_name) = allomorph_type_name(inner, mixins)
-                        && (type_name == "Allomorph" || type_name == allo_name)
-                    {
-                        return true;
-                    }
-                    return false;
-                }
-                if inner.isa_check(type_name) {
-                    return true;
-                }
-                // Also check mixin type keys (e.g., allomorphic "Str" mixin)
-                return mixins.contains_key(type_name);
-            }
-            Value::Proxy { .. } => "Proxy",
-            Value::ParametricRole { base_name, .. } => {
-                return base_name.resolve() == type_name;
-            }
-            Value::CustomType(c) => {
-                return c.name.resolve() == type_name;
-            }
-            Value::CustomTypeInstance(d) => {
-                return d.type_name.resolve() == type_name;
-            }
-            Value::Scalar(inner) => return inner.isa_check(type_name),
-            Value::ContainerRef(_) => return self.with_deref(|inner| inner.isa_check(type_name)),
-            Value::LazyThunk(thunk_data) => {
-                let cache = thunk_data.cache.lock().unwrap();
-                if let Some(ref cached) = *cache {
-                    return cached.isa_check(type_name);
-                }
-                "Scalar" // unforced lazy thunk
-            }
-            Value::LazyIoLines { .. } => "Seq",
-            Value::HashEntryRef { .. } => return self.hash_entry_read().isa_check(type_name),
-        };
-        if my_type == type_name {
-            return true;
-        }
-        // The X::Await::Died role is mixed into the original exception when
-        // `await` observes a broken Promise (see `await_died_error`): the cause
-        // keeps its own class but also does X::Await::Died.
-        if type_name == "X::Await::Died"
-            && let Value::Instance { attributes, .. } = self
-            && matches!(
-                attributes.as_map().get("__mutsu_does_await_died"),
-                Some(Value::Bool(true))
-            )
-        {
-            return true;
-        }
-        // Perl6::Metamodel:: and Metamodel:: are equivalent namespaces
-        if let Some(short) = my_type.strip_prefix("Perl6::")
-            && short == type_name
-        {
-            return true;
-        }
-        if let Some(short) = type_name.strip_prefix("Perl6::")
-            && short == my_type
-        {
-            return true;
-        }
-        // Check type hierarchy
-        match type_name {
-            "Any" => true,
-            "Mu" => true,
-            "SetHash" => matches!(self, Value::Set(_, true)),
-            "BagHash" => matches!(self, Value::Bag(_, true)),
-            "MixHash" => matches!(self, Value::Mix(_, true)),
-            "Cool" => {
-                matches!(
-                    self,
-                    Value::Int(_)
-                        | Value::BigInt(_)
-                        | Value::Num(_)
-                        | Value::Str(_)
-                        | Value::Bool(_)
-                        | Value::Rat(_, _)
-                        | Value::FatRat(_, _)
-                        | Value::BigRat(_, _)
-                        | Value::Complex(_, _)
-                        | Value::Array(..)
-                        | Value::Hash(_)
-                ) || matches!(
-                    self,
-                    Value::Instance { class_name, .. }
-                        if class_name == "Match" || class_name == "Capture"
-                )
-            }
-            "Capture" => {
-                matches!(self, Value::Capture { .. })
-                    || matches!(
-                        self,
-                        Value::Instance { class_name, .. }
-                            if class_name == "Match" || class_name == "Capture"
-                    )
-            }
-            "Numeric" => matches!(
-                self,
-                Value::Int(_)
-                    | Value::BigInt(_)
-                    | Value::Num(_)
-                    | Value::Rat(_, _)
-                    | Value::FatRat(_, _)
-                    | Value::BigRat(_, _)
-                    | Value::Complex(_, _)
-            ),
-            "Real" => matches!(
-                self,
-                Value::Int(_)
-                    | Value::BigInt(_)
-                    | Value::Num(_)
-                    | Value::Rat(_, _)
-                    | Value::FatRat(_, _)
-                    | Value::BigRat(_, _)
-            ),
-            "Rational" => matches!(
-                self,
-                Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-            ),
-            "Dateish" => matches!(
-                self,
-                Value::Instance { class_name, .. } if class_name == "Date" || class_name == "DateTime"
-            ),
-            "FatRat" => matches!(self, Value::FatRat(_, _) | Value::BigRat(_, _)),
-            "Int" => matches!(self, Value::Bool(_)),
-            "Stringy" => matches!(self, Value::Str(_)),
-            "Block" | "Routine" | "Code" | "Callable" => {
-                matches!(
-                    self,
-                    Value::Sub(_) | Value::WeakSub(_) | Value::Routine { .. }
-                ) || matches!(
-                    self,
-                    Value::Package(name)
-                        if matches!(name.resolve().as_str(), "Sub" | "Routine" | "Method" | "Block" | "Code")
-                )
-            }
-            "Method" => {
-                matches!(
-                    self,
-                    Value::Sub(data)
-                        if matches!(
-                            data.env.get("__mutsu_callable_type"),
-                            Some(Value::Str(kind)) if kind.as_str() == "Method"
-                        )
-                ) || matches!(
-                    self,
-                    Value::Instance { class_name, .. } if class_name == "Method"
-                ) || matches!(self, Value::Package(name) if name == "Method")
-            }
-            "Exception" => {
-                if let Value::Instance { class_name, .. } = self {
-                    class_name.resolve().starts_with("X::") || class_name == "Exception"
-                } else {
-                    false
-                }
-            }
-            "X::AdHoc" | "CX::Warn" | "CX::Return" | "X::OS" => {
-                if let Value::Instance { class_name, .. } = self {
-                    class_name == type_name
-                } else {
-                    false
-                }
-            }
-            "HyperSeq" => {
-                matches!(self, Value::HyperSeq(_))
-            }
-            "RaceSeq" => {
-                matches!(self, Value::RaceSeq(_))
-            }
-            "Seq" | "List" => {
-                matches!(
-                    self,
-                    Value::Array(..)
-                        | Value::LazyList(_)
-                        | Value::Slip(_)
-                        | Value::HyperSeq(_)
-                        | Value::RaceSeq(_)
-                )
-            }
-            "Positional" => {
-                matches!(
-                    self,
-                    Value::Array(..)
-                        | Value::LazyList(_)
-                        | Value::HyperSeq(_)
-                        | Value::RaceSeq(_)
-                        | Value::Range(_, _)
-                        | Value::RangeExcl(_, _)
-                        | Value::RangeExclStart(_, _)
-                        | Value::RangeExclBoth(_, _)
-                        | Value::GenericRange { .. }
-                        | Value::Capture { .. }
-                ) || matches!(
-                    self,
-                    Value::Package(name)
-                        if matches!(
-                            name.resolve().as_str(),
-                            "Array" | "List" | "Range" | "Buf" | "Blob" | "Capture"
-                        )
-                ) || matches!(
-                    self,
-                    Value::Instance { attributes, .. }
-                        if attributes.contains_key("__mutsu_array_storage")
-                )
-            }
-            "Map" | "Associative" => {
-                matches!(
-                    self,
-                    Value::Hash(_)
-                        | Value::Pair(_, _)
-                        | Value::ValuePair(_, _)
-                        | Value::Set(_, _)
-                        | Value::Bag(_, _)
-                        | Value::Mix(_, _)
-                        | Value::Capture { .. }
-                ) || matches!(
-                    self,
-                    Value::Package(name)
-                        if matches!(
-                            name.resolve().as_str(),
-                            "Hash" | "Map" | "Pair" | "Set" | "Bag" | "Mix" | "QuantHash" | "Capture"
-                        )
-                )
-            }
-            "Iterable" => matches!(
-                self,
-                Value::Array(..) | Value::LazyList(_) | Value::Hash(_) | Value::Seq(_)
-            ),
-            "ObjAt" => {
-                // ValueObjAt is a subclass of ObjAt
-                matches!(
-                    self,
-                    Value::Instance { class_name, .. }
-                        if class_name == "ObjAt" || class_name == "ValueObjAt"
-                )
-            }
-            "Pod::Block" => matches!(
-                self,
-                Value::Instance { class_name, .. }
-                    if class_name == "Pod::Block"
-                        || class_name == "Pod::Block::Comment"
-                        || class_name == "Pod::Block::Para"
-                        || class_name == "Pod::Block::Named"
-                        || class_name == "Pod::Heading"
-                        || class_name == "Pod::Block::Table"
-                        || class_name == "Pod::Item"
-            ),
-            "Pod::Config" => matches!(
-                self,
-                Value::Instance { class_name, .. }
-                    if class_name == "Pod::Config"
-            ),
-            _ => false,
-        }
-    }
-
-    /// Check if this value does (composes) the given role name.
-    pub(crate) fn does_check(&self, role_name: &str) -> bool {
-        if let Value::Mixin(inner, mixins) = self {
-            let key = format!("__mutsu_role__{}", role_name);
-            if mixins.contains_key(&key) {
-                return true;
-            }
-            return inner.does_check(role_name);
-        }
-        // Check built-in role compositions
-        if role_name == "Encoding" {
-            if let Value::Instance { class_name, .. } = self
-                && class_name == "Encoding::Builtin"
-            {
-                return true;
-            }
-            if let Value::Package(name) = self
-                && name == "Encoding::Builtin"
-            {
-                return true;
-            }
-        }
-        // Delegate to isa_check for other cases (roles are stored as parents)
-        self.isa_check(role_name)
-    }
-}
-
 /// Returns the Raku type name for a value (used in error messages).
 pub(crate) fn what_type_name(val: &Value) -> String {
     match val {
